third_party/WebKit/LayoutTests/inspector/css_tracker/css-tracker-view.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>

<link rel="stylesheet" type="text/css" href="resources/highlight-in-source.css">

<script>

var initialize_CSSTracker = function() {
    InspectorTest.preloadModule("css_tracker");
}

function test()
{
    var tracker = self.runtime.sharedInstance(CSSTracker.CSSTrackerView);
    InspectorTest.addSniffer(tracker._listView, "update", displayResults);

    tracker._toggleRecording(true);
    UI.viewManager.showView("css_tracker");
    tracker._toggleRecording(false);

    function displayResults()
    {
        var dataGrid = tracker._listView._dataGrid;
        for (var child of dataGrid.rootNode().children) {
            var data = child._coverageInfo;
            var url = InspectorTest.formatters.formatAsURL(data.url);
            InspectorTest.addResult(`${url} used: ${data.usedSize} unused: ${data.unusedSize} total: ${data.size}`);
        }
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p class="class">
Tests the unused CSS list after finishing recording in the CSSTracker view.
</p>

</body>
</html>
